Split negative flash into 2 files
